requirements.txt
pyyaml>=6.0
python-dateutil>=2.8.2
matplotlib>=3.6.0
pandas>=1.5.0
numpy>=1.23.0
